initial service commands implementation #1 #3
